installer: move the source invariants into one integration test
